Move PsfFont to separate header. Add some predefined colors.

+++ kernel/arch/x86_64/memory.cpp
@@ -2,7 +2,7 @@
 
 /*** DO NOT USE THESE AFTER INITIALIZATION ***/
 extern const void *_kernel_start;    // Start of kernel in physical address space
-extern const uint64_t _kernel_pages; // Number of contiguous pages from start of kernel
+extern const uint32_t _kernel_pages; // Number of contiguous pages from start of kernel
 extern const BrewOS::Memory::MemoryMap *_memory_map;
 
 namespace BrewOS::Memory
